requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.11.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.6.2
apispec==6.3.0
appnope==0.1.3
argilla==1.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arxiv==1.4.8
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
awscli==1.32.103
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
bitarray==2.7.3
black==23.3.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
botocore==1.34.103
cached-property==1.5.2
cachetools==5.3.1
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.0.12
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.23
colorama==0.4.3
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
coverage==7.2.7
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duckdb==0.7.1
EB==0.1.5
ebcli==4.0.18
email-validator==1.1.3
entrypoints==0.4
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.1
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.9.0
flake8==6.0.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Admin==1.6.1
-e git+ssh://git@github.com/appasahebs/wp-eks.git@9bb5ea8e5fbc87f78528368451af176db8039ba7#egg=flask_api_tutorial&subdirectory=flask-api-tutorial
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
flask-restx==1.1.0
Flask-Script==2.0.6
flask-smorest==0.42.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
grpcio==1.49.1
h11==0.14.0
hnswlib==0.7.0
httpcore==1.0.7
httptools==0.5.0
httpx==0.27.2
huggingface-hub==0.12.0
identify==2.5.24
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
install==1.3.5
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.1
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.10.3
jupyterlab-widgets==3.0.8
jwt==1.3.1
keyring==9.0
keyrings.alt==4.2.0
kiwisolver==1.4.4
langchain==0.0.178
lazy-object-proxy==1.9.0
lxml==4.9.2
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
ollama==0.4.6
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pdfreader==0.1.12
pexpect==4.8.0
pgvector==0.1.8
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.2.0
posthog==3.0.1
pprintpp==0.4.0
pre-commit==3.3.3
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==1.4.2
pylint==2.17.4
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyproject-api==1.5.3
pyrsistent==0.19.3
pytest==7.4.0
pytest-black==0.3.12
pytest-clarity==1.0.1
pytest-cov==4.1.0
pytest-dotenv==0.5.2
pytest-flake8==1.1.1
pytest-flask==1.2.0
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.6
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.3
QtPy==2.3.1
ray==2.4.0
regex==2022.10.31
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.7.2
s3transfer==0.10.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.8.5
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.14.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4
SQLAlchemy==1.4.23
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
termcolor==1.1.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tox==4.6.4
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.12.2
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.19.0
voila==0.4.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.1.9
webargs==8.2.0
webcolors==1.12
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.4
Werkzeug==2.3.4
widgetsnbextension==4.0.8
wikipedia==1.4.0
wrapt==1.14.1
WTForms==3.0.1
XlsxWriter==3.1.0
y-py==0.5.5
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.13.0
zstandard==0.21.0
